requirements.txt
tqdm
pillow
tensorflow
pandas
numpy